Testing KDTree still

+++ pearce/emulator/emu.py
@@ -1131,7 +1131,7 @@ class ExtraCrispy(Emu):
             # If points cannot be evenly divided, there'll be some skipped ones.
             # We'll add them in at the end.
             n_missed = np.sum([len(leaf)%self.experts for leaf in leaves])
-            missed_points = np.zeros(n_missed)
+            missed_points = np.zeros(n_missed, dtype=int)
             missed_idx = 0
 
             for i, leaf in enumerate(leaves):
